+++ src/Asset/MceAssets.php
@@ -1,16 +1,12 @@
 <?php
-/**
- * Contao Open Source CMS
- *
+
+/*
  * Copyright (c) 2020 Heimrich & Hannot GmbH
  *
- * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
- * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ * @license LGPL-3.0-or-later
  */
 
-
 namespace HeimrichHannot\MultiColumnEditorBundle\Asset;
-
 
 use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 
@@ -33,9 +29,6 @@ class MceAssets
         $this->containerUtil = $containerUtil;
     }
 
-    /**
-     * @param \HeimrichHannot\EncoreBundle\Asset\FrontendAsset $encoreFrontendAsset
-     */
     public function setEncoreFrontendAsset(\HeimrichHannot\EncoreBundle\Asset\FrontendAsset $encoreFrontendAsset): void
     {
         $this->encoreFrontendAsset = $encoreFrontendAsset;
@@ -47,6 +40,7 @@ class MceAssets
             $GLOBALS['TL_CSS']['contao-multi-column-editor-bundle'] = 'bundles/heimrichhannotcontaomulticolumneditor/contao-multi-column-editor-bundle-be.css|static';
         } else {
             $GLOBALS['TL_JAVASCRIPT']['sortablejs'] = 'assets/sortablejs/sortablejs/Sortable.min.js|static';
+
             if ($this->encoreFrontendAsset) {
                 $this->encoreFrontendAsset->addActiveEntrypoint('contao-multi-column-editor-bundle');
             }