+++ src/Resources/assets/js/contao-multi-column-editor-bundle.js
@@ -1,1 +1,1 @@
-import '@hundh/contao-multi-column-editor-bundle';
+import '../../npm-package/js/contao-multi-column-editor-bundle';